REFACTORING : entity list 초기화
