+++ src/main/java/gift/exception/optionException/OptionException.java
@@ -1,4 +1,4 @@
-package gift.exception;
+package gift.exception.optionException;
 
 public class OptionException extends RuntimeException{
     public OptionException(String message) {
